refactor: add vite-plugin-pages
