--- demos/vue-vite/src/components/Playground.vue
@@ -1,3 +1,0 @@
-<script setup lang="ts">
-const props = defineProps<{ name: string; age: number }>()
-</script>
