--- demos/vue-vite/src/components/SocialCard.vue
@@ -1,13 +1,0 @@
-<script setup lang="ts">
-defineProps<{
-  twitter?: String
-  facebook?: String
-  linkedin?: String
-  email?: String
-  youtube?: String
-}>()
-</script>
-
-<template>
-  {{ twitter }}
-</template>
