--- demos/vue-vite/src/composables/useCounter.ts
@@ -1,15 +1,0 @@
-export function useCounter() {
-  const count = ref(0)
-
-  const increment = () => {
-    count.value++
-  }
-
-  const double = computed(() => count.value * 2)
-
-  return {
-    count,
-    double,
-    increment
-  }
-}
