+++ demos/vue-vite/src/pages/index.vue
