+++ demos/vue-vite/src/pages/transition/components/FadeTransition.vue
