+++ demos/vue-vite/src/pages/vue-3-3-playground/components/SimpleText.vue
