+++ demos/vue-vite/src/pages/writeable-computed/index.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Parent from '@/components/writable-computed-v-model/Parent.vue'
+import Parent from './components/Parent.vue'
 </script>
 
 <template>
