+++ demos/vue-vite/src/vite-env.d.ts
@@ -1,4 +1,5 @@
 /// <reference types="vite/client" />
+/// <reference types="vite-plugin-pages/client" />
 
 declare module '*.vue' {
   import type { DefineComponent } from 'vue'
