+++ demos/vue-vite/vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import Icons from 'unplugin-icons/vite'
+import Pages from 'vite-plugin-pages'
 
 import { resolve } from 'path'
 // https://vitejs.dev/config/
@@ -10,6 +11,11 @@ export default defineConfig({
       '@': resolve(__dirname, './src')
     }
   },
-
-  plugins: [Icons({ compiler: 'vue3', autoInstall: true }), vue({})]
+  plugins: [
+    Icons({ compiler: 'vue3', autoInstall: true }),
+    vue({}),
+    Pages({
+      exclude: ['**/components/*.vue']
+    })
+  ]
 })
